fixed the logic and add .ignore

+++ Forward/Motors/IEngine.cs
@@ -34,8 +34,8 @@ namespace Forward.Motors {
         /// Коэффициент зависимости скорости охлаждения от температуры двигателя и окружающей среды.
         /// </summary>
         public double C { get; set; }
-        
-        public int TimeWork { get; set; }
+
+        public double TimeWork { get; set; } 
         public double TemperatureEnvironment { get; set; }
         public double TemperatureEngine { get; set; }   
     }
